+++ crates/cheatcodes/assets/cheatcodes.json
@@ -6256,26 +6256,6 @@
     },
     {
       "func": {
-        "id": "getPvmInfo",
-        "description": "Gets PVM-specific information for the current execution context.\nReturns information about PVM state and mode.\nExample usage:\nstring memory info = vm.getPvmInfo();",
-        "declaration": "function getPvmInfo() external view returns (string memory);",
-        "visibility": "external",
-        "mutability": "view",
-        "signature": "getPvmInfo()",
-        "selector": "0x71939732",
-        "selectorBytes": [
-          113,
-          147,
-          151,
-          50
-        ]
-      },
-      "group": "utilities",
-      "status": "stable",
-      "safety": "safe"
-    },
-    {
-      "func": {
         "id": "getRecordedLogs",
         "description": "Gets all the recorded logs.",
         "declaration": "function getRecordedLogs() external returns (Log[] memory logs);",
@@ -8159,7 +8139,7 @@
     {
       "func": {
         "id": "pvm",
-        "description": "Enables or disables PVM (PolkaVM) mode for contract execution.\nWhen enabled, contracts will be executed using the PVM runtime instead of EVM.\nExample usage:\nvm.pvm(true);  // Enable PVM mode\nvm.pvm(false); // Disable PVM mode (use EVM)",
+        "description": "Enables or disables PVM (PolkaVM) mode for contract execution.\nWhen enabled, only PVM-specific calls are intercepted; all other operations run on EVM.\nExample usage:\nvm.pvm(true);  // Enable PVM mode\nvm.pvm(false); // Disable PVM mode (use EVM)",
         "declaration": "function pvm(bool enabled) external;",
         "visibility": "external",
         "mutability": "",
