+++ crates/cheatcodes/spec/src/vm.rs
@@ -2879,19 +2879,12 @@ interface Vm {
     function interceptInitcode() external;
 
     /// Enables or disables PVM (PolkaVM) mode for contract execution.
-    /// When enabled, contracts will be executed using the PVM runtime instead of EVM.
+    /// When enabled, only PVM-specific calls are intercepted; all other operations run on EVM.
     /// Example usage:
     /// vm.pvm(true);  // Enable PVM mode
     /// vm.pvm(false); // Disable PVM mode (use EVM)
     #[cheatcode(group = Utilities)]
     function pvm(bool enabled) external;
-
-    /// Gets PVM-specific information for the current execution context.
-    /// Returns information about PVM state and mode.
-    /// Example usage:
-    /// string memory info = vm.getPvmInfo();
-    #[cheatcode(group = Utilities)]
-    function getPvmInfo() external view returns (string memory);
 }
 }
